+++ repo/installer.go
@@ -18,7 +18,6 @@ import (
 	"github.com/Khulnasoft-lab/gopkg/msg"
 	gpath "github.com/Khulnasoft-lab/gopkg/path"
 	"github.com/Khulnasoft-lab/gopkg/util"
-	"github.com/Khulnasoft-lab/semver"
 	"github.com/Khulnasoft-lab/vcs"
 	"github.com/codegangsta/cli"
 )