+++ repo/vcs.go
@@ -16,7 +16,6 @@ import (
 	"github.com/Khulnasoft-lab/gopkg/cfg"
 	"github.com/Khulnasoft-lab/gopkg/msg"
 	gpath "github.com/Khulnasoft-lab/gopkg/path"
-	"github.com/Khulnasoft-lab/semver"
 	v "github.com/Khulnasoft-lab/vcs"
 )